# openapi.yaml
openapi: 3.0.3
info:
  title: MIT Open API
  version: 0.0.1
  description: MIT public API
paths:
...
  /api/v1/learning_resources/:
    get:
      operationId: learning_resources_list
      description: Get a paginated list of learning resources.
      summary: List
      parameters:
      - in: query
        name: department
        schema:
          type: string
          enum:
          - '1'
          - '10'
          - '11'
          - '12'
          - '14'
          - '15'
          - '16'
          - '17'
          - '18'
          - '2'
          - '20'
          - 21A
          - 21G
          - 21H
          - 21L
          - 21M
          - '22'
          - '24'
          - '3'
          - '4'
          - '5'
          - '6'
          - '7'
          - '8'
          - '9'
          - CC
          - CMS-W
          - EC
          - ES
          - ESD
          - HST
          - IDS
          - MAS
          - PE
          - RES
          - STS
          - WGS
        description: |-
          Department ID

          * `1` - Civil and Environmental Engineering
          * `2` - Mechanical Engineering
          * `3` - Materials Science and Engineering
          * `4` - Architecture
          * `5` - Chemistry
          * `6` - Electrical Engineering and Computer Science
          * `7` - Biology
          * `8` - Physics
          * `9` - Brain and Cognitive Sciences
          * `10` - Chemical Engineering
          * `11` - Urban Studies and Planning
          * `12` - Earth, Atmospheric, and Planetary Sciences
          * `14` - Economics
          * `15` - Sloan School of Management
          * `16` - Aeronautics and Astronautics
          * `17` - Political Science
          * `18` - Mathematics
          * `20` - Biological Engineering
          * `21A` - Anthropology
          * `21G` - Global Studies and Languages
          * `21H` - History
          * `21L` - Literature
          * `21M` - Music and Theater Arts
          * `22` - Nuclear Science and Engineering
          * `24` - Linguistics and Philosophy
          * `CC` - Concourse
          * `CMS-W` - Comparative Media Studies/Writing
          * `EC` - Edgerton Center
          * `ES` - Experimental Study Group
          * `ESD` - Engineering Systems Division
          * `HST` - Health Sciences and Technology
          * `IDS` - Institute for Data, Systems, and Society
          * `MAS` - Media Arts and Sciences
          * `PE` - Athletics, Physical Education and Recreation
          * `RES` - Supplemental Resources
          * `STS` - Science, Technology, and Society
          * `WGS` - Women's and Gender Studies
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: offered_by
        schema:
          type: string
          enum:
          - bootcamps
          - csail
          - ctl
          - mitpe
          - mitx
          - ocw
          - scc
          - see
          - xpro
        description: |-
          Offered By

          * `mitx` - MITx
          * `ocw` - OCW
          * `bootcamps` - Bootcamps
          * `xpro` - xPRO
          * `csail` - CSAIL
          * `mitpe` - Professional Education
          * `see` - Sloan Executive Education
          * `scc` - Schwarzman College of Computing
          * `ctl` - Center for Transportation & Logistics
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: query
        name: platform
        schema:
          type: string
          enum:
          - bootcamps
          - csail
          - ctl
          - edx
          - emeritus
          - globalalumni
          - mitpe
          - mitxonline
          - ocw
          - oll
          - podcast
          - scc
          - see
          - simplilearn
          - susskind
          - whu
          - xpro
        description: |-
          Platform

          * `edx` - edX
          * `ocw` - OCW
          * `oll` - Open Learning Library
          * `mitxonline` - MITx Online
          * `bootcamps` - Bootcamps
          * `xpro` - xPRO
          * `csail` - CSAIL
          * `mitpe` - Professional Education
          * `see` - Sloan Executive Education
          * `scc` - Schwarzman College of Computing
          * `ctl` - Center for Transportation & Logistics
          * `whu` - WHU
          * `susskind` - Susskind
          * `globalalumni` - Global Alumni
          * `simplilearn` - Simplilearn
          * `emeritus` - Emeritus
          * `podcast` - Podcast
      - in: query
        name: professional
        schema:
          type: boolean
      - in: query
        name: resource_type
        schema:
          type: string
          enum:
          - course
          - learning_path
          - podcast
          - podcast_episode
          - program
        description: |-
          Resource Type

          * `course` - Course
          * `program` - Program
          * `learning_path` - Learning Path
          * `podcast` - Podcast
          * `podcast_episode` - Podcast Episode
      tags:
      - learning_resources
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLearningResourceList'
          description: ''
...
  /api/v1/learning_resources/{id}/:
    get:
      operationId: learning_resources_retrieve
      description: Retrieve a single learning resource.
      summary: Retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this learning resource.
        required: true
      tags:
      - learning_resources
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningResource'
          description: ''
...
  /api/v1/learningpaths/{parent_id}/resources/:
    get:
      operationId: learningpaths_resources_list
      description: Viewset for LearningPath related resources
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: path
        name: parent_id
        schema:
          type: integer
        required: true
      tags:
      - learningpaths
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLearningPathRelationshipList'
          description: ''
    post:
      operationId: learningpaths_resources_create
      description: Viewset for LearningPath related resources
      parameters:
      - in: path
        name: parent_id
        schema:
          type: integer
        required: true
      tags:
      - learningpaths
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningPathRelationshipRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LearningPathRelationshipRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LearningPathRelationshipRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPathRelationship'
          description: ''
  /api/v1/learningpaths/{parent_id}/resources/{id}/:
    get:
      operationId: learningpaths_resources_retrieve
      description: Viewset for LearningPath related resources
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this learning resource relationship.
        required: true
      - in: path
        name: parent_id
        schema:
          type: integer
        required: true
      tags:
      - learningpaths
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPathRelationship'
          description: ''
    put:
      operationId: learningpaths_resources_update
      description: Viewset for LearningPath related resources
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this learning resource relationship.
        required: true
      - in: path
        name: parent_id
        schema:
          type: integer
        required: true
      tags:
      - learningpaths
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningPathRelationshipRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LearningPathRelationshipRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LearningPathRelationshipRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPathRelationship'
          description: ''
    patch:
      operationId: learningpaths_resources_partial_update
      description: Viewset for LearningPath related resources
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this learning resource relationship.
        required: true
      - in: path
        name: parent_id
        schema:
          type: integer
        required: true
      tags:
      - learningpaths
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedLearningPathRelationshipRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedLearningPathRelationshipRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedLearningPathRelationshipRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPathRelationship'
          description: ''
    delete:
      operationId: learningpaths_resources_destroy
      description: Viewset for LearningPath related resources
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this learning resource relationship.
        required: true
      - in: path
        name: parent_id
        schema:
          type: integer
        required: true
      tags:
      - learningpaths
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/userlists/{parent_id}/resources/:
    get:
      operationId: userlists_resources_list
      description: Viewset for UserListRelationships
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - in: path
        name: parent_id
        schema:
          type: integer
        required: true
      tags:
      - userlists
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserListRelationshipList'
          description: ''
    post:
      operationId: userlists_resources_create
      description: Viewset for UserListRelationships
      parameters:
      - in: path
        name: parent_id
        schema:
          type: integer
        required: true
      tags:
      - userlists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserListRelationshipRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserListRelationshipRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserListRelationshipRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListRelationship'
          description: ''
...
components:
  schemas:
...
    LearningPathRelationship:
      type: object
      description: Specialized serializer for a LearningPath relationship
      properties:
        id:
          type: integer
          readOnly: true
        resource:
          allOf:
          - $ref: '#/components/schemas/LearningResource'
          readOnly: true
        position:
          type: integer
          maximum: 2147483647
          minimum: 0
        parent:
          type: integer
        child:
          type: integer
      required:
      - child
      - id
      - parent
      - resource
    LearningPathRelationshipRequest:
      type: object
      description: Specialized serializer for a LearningPath relationship
      properties:
        position:
          type: integer
          maximum: 2147483647
          minimum: 0
        parent:
          type: integer
        child:
          type: integer
      required:
      - child
      - parent
...
    LearningResource:
      oneOf:
      - $ref: '#/components/schemas/ProgramResource'
      - $ref: '#/components/schemas/CourseResource'
      - $ref: '#/components/schemas/LearningPathResource'
      - $ref: '#/components/schemas/PodcastResource'
      - $ref: '#/components/schemas/PodcastEpisodeResource'
      discriminator:
        propertyName: resource_type
        mapping:
          program: '#/components/schemas/ProgramResource'
          course: '#/components/schemas/CourseResource'
          learning_path: '#/components/schemas/LearningPathResource'
          podcast: '#/components/schemas/PodcastResource'
          podcast_episode: '#/components/schemas/PodcastEpisodeResource'
...
    PaginatedLearningPathRelationshipList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/LearningPathRelationship'
...
    PaginatedLearningResourceList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/LearningResource'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid